+++ tests/fixtures/deep_config_file.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Config file with deeply nested test data for the unit tests.
+ *
+ * @package   BrightNucleus\Config
+ * @author    Alain Schlesser <alain.schlesser@gmail.com>
+ * @license   GPL-2.0+
+ * @link      http://www.brightnucleus.com/
+ * @copyright 2016 Alain Schlesser, Bright Nucleus
+ */
+
+namespace BrightNucleus\Config;
+
+$test_data = [
+    'section_1' => [
+        'test_key_1' => 'test_value_1',
+    ],
+    'section_2' => [
+        'test_key_2' => 'test_value_2',
+    ],
+];
+
+return ['vendor' => ['package' => $test_data]];